move 'curry' tests to separate file
